rename to config.tempo to config
