+++ main.go
@@ -15,7 +15,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	configtempov1alpha1 "github.com/os-observability/tempo-operator/apis/config.tempo/v1alpha1"
+	configtempov1alpha1 "github.com/os-observability/tempo-operator/apis/config/v1alpha1"
 	tempov1alpha1 "github.com/os-observability/tempo-operator/apis/tempo/v1alpha1"
 	controllers "github.com/os-observability/tempo-operator/controllers/tempo"
 	//+kubebuilder:scaffold:imports
